+++ src/main.cpp
@@ -53,6 +53,30 @@ $on_mod(Loaded) {
         }
     });
 
+    listenForSettingChanges<bool>("show-accuracy", [](bool) {
+        if (auto pl = static_cast<ProPlayLayer*>(PlayLayer::get())) {
+            if (auto overlay = pl->m_fields->overlay) {
+                if (pl->m_fields->started) {
+                    queueInMainThread([overlay = Ref(overlay)] {
+                        overlay->LABELS();
+                    });
+                }
+            }
+        }
+    });
+
+    listenForSettingChanges<bool>("show-combo", [](bool) {
+        if (auto pl = static_cast<ProPlayLayer*>(PlayLayer::get())) {
+            if (auto overlay = pl->m_fields->overlay) {
+                if (pl->m_fields->started) {
+                    queueInMainThread([overlay = Ref(overlay)] {
+                        overlay->LABELS();
+                    });
+                }
+            }
+        }
+    });
+
 }
 
 class $modify(PauseLayer) {