+++ changelog.md
@@ -1,3 +1,9 @@
+# v1.2.0
+
+* Added accuracy %
+* Added combo counter
+* Added input judgements
+
 # v1.1.0
 
 * Added circle and square note styles